requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.0.12
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
idna==3.4
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
requests==2.26.0
sqlparse==0.4.3
urllib3==1.26.15
